refactor: :rainbow: 04.Models 불필요 생성자 삭제 및 Provider 그룹화
