+++ iOS/IssueTracker/IssueTracker/04.Models/Provider/IssueFilter.swift
